Telescope_info table split into Telescope_info and Array_info tables.
Telescope_info now features a row per telescope type, containing focal length,
number of pixels, and pixel positions. To circunvent the issue of storing
arrays of variable length for the positions the pixel_pos placeholder's
shape corresponds to the camera with the largest number of pixels in the
among the selected telescopes, and for cameras with a smaller number of pixels
a value of 0 is stored in pixel_pos as the coordinates for unexisting pixels.
This is hopefully a temporary fix.

The new Array_info table contains a row per telescope in the array, showing
telescope id, position in the array, and pointing position.