Telescope_info table split into Telescope_info and Array_info tables.
Telescope_info now features a row per telescope type, containing focal length,
number of pixels, and pixel positions. To circunvent the issue of storing
arrays of variable length for the positions the pixel_pos placeholder's
shape corresponds to the camera with the largest number of pixels in the
among the selected telescopes, and for cameras with a smaller number of pixels
a value of 0 is stored in pixel_pos as the coordinates for unexisting pixels.
This is hopefully a temporary fix.

The new Array_info table contains a row per telescope in the array, showing
telescope id, position in the array, and pointing position.


Former-commit-id: 58d52e9cc3e1840d8bc443f47907d26c2f294110 [formerly 58d52e9cc3e1840d8bc443f47907d26c2f294110 [formerly a243d1f0799fcae9e4a49486aebc20d124f21a8e]]
Former-commit-id: e3609888855f1a9946b107e3ce7ad65febe67648
Former-commit-id: 2fc72a1bd6bdadb67000b0ecf364c8b5f98b377e