Make example description and metadata general
The metadata dictionary as used in CTLearn has been eliminated and
replaced with an "example description", which is a list of dictionaries
that describe the example array at the corresponding index. These
contain the shape and dtype information needed to load the data. The
number of total and selected telescopes are available as class
attributes. To get the number of examples, a function is provided that
calculates the number of examples grouped by any parameter included in
the example. This general approach allows calculation of the number of
examples divided arbitrarily, for example, by gamma/hadron label, energy
bin, or multiplicity, as needed. However, as these correspond to the
selected events, getting the number of total pre-selection-cut examples
as needed to calculate q-factor will require the construction of a
second DL1DataReader.


Former-commit-id: 6d7e0a1cc914e2dba5c80f0126b42ec137fc95e8