Read array info using the tel_type and tel_id
Fixes a bug in which, when just the tel_id is used, array info from two
telescopes of different tel_types is read, because the tel_ids are the
same. This is due to an overflow bug, present only in legacy files, in
which the tel_id numbering after 255 resets to 0.