installation fix with removing ctapipe from setup.py
Since the DL1 writing will be done in ctapipe in the future, having ctapipe as a dependency is not required anymore. For the reader (plus additional classes), we only need a light installation. If users want to write DL1 data with the DL1DataWriter, they are advice to follow the instruction for a full installation.

Update README accordingly.