Make transform compatible with single-task use
For single task use, it makes sense to treat the particle type as a
class label, as it is called throughout CTLearn. To preserve
compatibility with both single- and multi-task, make the transformed
particle type name a configurable parameter. To maintain the current
functionality, pass 'particletype' as an argument to the transform.