Make HDF5 access threading safe
Wrap all reads of the HDF5 files in a threading lock, to support
parallel reads. While this had been previously done at the time of
opening the files, this was not sufficient because we keep all files
open for the lifetime of the reader for performance, and later accesses
were not thread protected.