modified the transformations
Use an argument to specify the name of the transform class.

Set the default arguments of the transforms to the default settings of the multitask branch in CTLearn. (The GammaLearn team has it's own transformation)

For some transforms an extra dimension is added to directly use the output as labels in CTLearn.