Update reader for the new format
add "image_selection_from_file" parameter. It is a dict where one cut can be stored (at the moment).

If "image_selection_from_file" is present "image_selection" must not be present because the first one is for the new format of h5 file, the second one for the old format.