reconstruct the SrcPosX/Y in the camera for the arrival direction regression in MAGIC
As discussed in the call, it is better to reconstruct the SrcPosX/Y in the camera for the arrival direction regression. This is a dirty fix for the time being here. We need to come up with a more appropriate solution in ctapipe_io_magic for this regression task (under discussion).