added ability to read from files with split_datasets_by tel_type
It will automatically detect the split, which was used in the ctapipe-stage1/ctapipe-merge tool and deal with the different 'Group' names. The table itself have the same structure.

get rid of some local variables that occupied some memory

store unshuffled example_identifiers to disk and shuffle them afterwards.

minor bug fix -> read *_TRANSFORM_SCALE from the first table and not from the table tel_001. It broke before, when LST1 is not in the file.