Fix error from unneeded memory use
Initializing a DL1DataReaderSTAGE1 on my machine results in a crash with the following error message:

MemoryError: Unable to allocate 23.4 GiB for an array with shape (262144, 120, 200) and data type float32

The cause is a few lines that are computing num_showers using shower histograms in the simulation files. In fact, num_showers is never subsequently used in reader.py (n_showers is used instead). Deleting these lines solves the error.